ST: fix current and add new tests for multiple namespace (#1078)
* Added new functionality for improve CO deployment (now it can watch multiple namespaces). 
* Added few new tests for watching multiple namespaces
