+++ api/src/test/java/io/strimzi/test/StrimziExtension.java
@@ -49,6 +49,7 @@ import java.util.HashSet;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.Stack;
 import java.util.function.Consumer;
 import java.util.function.Supplier;
@@ -265,7 +266,7 @@ public class StrimziExtension implements AfterAllCallback, BeforeAllCallback, Af
     }
 
     /**
-     * TODO
+     * Class for annotations routine execution
      */
     abstract class Bracket extends Statement implements Runnable {
         public final Statement statement;
@@ -661,6 +662,9 @@ public class StrimziExtension implements AfterAllCallback, BeforeAllCallback, Af
                     LOGGER.info("Creating namespace '{}' before test per @Namespace annotation on {}", namespace.value(), name(element));
                     kubeClient().createNamespace(namespace.value());
                     previousNamespace = namespace.use() ? kubeClient().namespace(namespace.value()) : kubeClient().namespace();
+                    if (element instanceof Method) {
+                        applyMultipleNamespacesWatcher(element);
+                    }
                 }
 
                 @Override
@@ -674,9 +678,39 @@ public class StrimziExtension implements AfterAllCallback, BeforeAllCallback, Af
         return last;
     }
 
+    private void applyMultipleNamespacesWatcher(AnnotatedElement element) {
+        List<Namespace> namespaces = annotations(element, Namespace.class);
+        String defaultNamespace  = namespaces.get(0).value();
+        for (Namespace namespace: namespaces) {
+            if (namespace.value().matches(defaultNamespace)) {
+                continue;
+            }
+            Map<File, String> configYamlFiles = Arrays.stream(
+                    Objects.requireNonNull(new File(CO_INSTALL_DIR).listFiles((file, name) -> name.matches("[0-9]*-RoleBinding.*")))
+            ).sorted().collect(Collectors.toMap(file -> file, f -> getContent(f, node -> {
+
+                ArrayNode subjects = (ArrayNode) node.get("subjects");
+                ObjectNode subject = (ObjectNode) subjects.get(0);
+                subject.put("kind", "ServiceAccount")
+                        .put("name", "strimzi-cluster-operator")
+                        .put("namespace", defaultNamespace);
+
+            }), (x, y) -> x, LinkedHashMap::new));
+
+            for (Map.Entry<File, String> entry : configYamlFiles.entrySet()) {
+                LOGGER.info("Apply {} into namespace {}", entry.getKey(), namespace.value());
+                kubeClient().namespace(namespace.value());
+                kubeClient().clientWithAdmin().applyContent(entry.getValue());
+            }
+        }
+        kubeClient().namespace(defaultNamespace);
+    }
+
     @SuppressWarnings("unchecked")
     private Statement installOperatorFromExamples(AnnotatedElement element, Statement last, ClusterOperator cc) {
-        Map<File, String> yamls = Arrays.stream(new File(CO_INSTALL_DIR).listFiles()).sorted().collect(Collectors.toMap(file -> file, f -> getContent(f, node -> {
+        Map<File, String> yamls = Arrays.stream(
+                Objects.requireNonNull(new File(CO_INSTALL_DIR).listFiles())
+        ).sorted().collect(Collectors.toMap(file -> file, f -> getContent(f, node -> {
             // Change the docker org of the images in the 04-deployment.yaml
             if ("050-Deployment-strimzi-cluster-operator.yaml".equals(f.getName())) {
                 ObjectNode containerNode = (ObjectNode) node.get("spec").get("template").get("spec").get("containers").get(0);
@@ -692,7 +726,7 @@ public class StrimziExtension implements AfterAllCallback, BeforeAllCallback, Af
                 containerNode.replace("resources", resources);
                 containerNode.remove("resources");
                 JsonNode ccImageNode = containerNode.get("image");
-                ((ObjectNode) containerNode).put("image", TestUtils.changeOrgAndTag(ccImageNode.asText()));
+                containerNode.put("image", TestUtils.changeOrgAndTag(ccImageNode.asText()));
                 for (JsonNode envVar : containerNode.get("env")) {
                     String varName = envVar.get("name").textValue();
                     // Replace all the default images with ones from the $DOCKER_ORG org and with the $DOCKER_TAG tag
@@ -710,6 +744,16 @@ public class StrimziExtension implements AfterAllCallback, BeforeAllCallback, Af
                         if (varName.equals(envVariable.key())) {
                             ((ObjectNode) envVar).put("value", envVariable.value());
                         }
+                    }
+
+                    if (varName.matches("STRIMZI_NAMESPACE")) {
+                        List<Namespace> namespaces = annotations(element, Namespace.class);
+                        List<String> test = new ArrayList<>();
+                        ((ObjectNode) envVar).remove("valueFrom");
+                        for (Namespace namespace : namespaces) {
+                            test.add(namespace.value());
+                        }
+                        ((ObjectNode) envVar).put("value", String.join(",", test));
                     }
                 }
             }
@@ -735,9 +779,11 @@ public class StrimziExtension implements AfterAllCallback, BeforeAllCallback, Af
                 for (Map.Entry<File, String> entry : yamls.entrySet()) {
                     LOGGER.info("creating possibly modified version of {}", entry.getKey());
                     deletable.push(entry.getValue());
+
                     kubeClient().namespace(annotations(element, Namespace.class).get(0).value());
                     kubeClient().clientWithAdmin().applyContent(entry.getValue());
                 }
+                applyMultipleNamespacesWatcher(element);
                 kubeClient().waitForDeployment(CO_DEPLOYMENT_NAME, 1);
             }
 
@@ -769,7 +815,7 @@ public class StrimziExtension implements AfterAllCallback, BeforeAllCallback, Af
                 entry("logLevel", OPERATOR_LOG_LEVEL))
                 .collect(entriesToMap()));
 
-        /** These entries aren't applied to the deployment yaml at this time */
+        /* These entries aren't applied to the deployment yaml at this time */
         Map<String, String> envVars = Collections.unmodifiableMap(Arrays.stream(cc.envVariables())
                 .map(var -> entry(String.format("env.%s", var.key()), var.value()))
                 .collect(entriesToMap()));
